==> notebooks/01-metadata.ipynb <==
# Video Metadata

The first step is to figure out how to capture information about Sydney's videos for the month of January 2023. This notebook shows my process.

Note: Section links work in JupyterLab but do not work when viewed in Github.

1. [Request](#Request-a-response) a response from YouTube's API
2. [Create](#Create-a-dataframe-with-channel-metadata) a dataframe with channel metadata
3. [Augment](#Augment-dataframe-with-video-metadata) dataframe with video metadata
4. [Save](#Save-the-data) the data to use later

### Request a response

* Load up your API key into memory. My way is below. You may have a different way to do this.

with open('../config/youtube_api') as file:
    DEVELOPER_KEY = file.readline()

* The next two cells get you set up to make requests with the YouTube API. 

# API client library
import googleapiclient.discovery
# API information
api_service_name = "youtube"
api_version = "v3"

# API client
youtube = googleapiclient.discovery.build(
    api_service_name, api_version, developerKey = DEVELOPER_KEY)

* The next cell specifies the YouTube channel id. Perhaps this will be obvious to others, but it took me a while to figure out that this value is NOT the name of the YouTube channel, but the hash that goes along with it. It took me a while to track it down.

* If you want to query some other channel's data, take a break and go figure out what its `channelId` is. 

CHANNEL_ID = 'UCVQJZE_on7It_pEv6tn-jdA' # Sydney's channel id

* I wanted to see what a request to the youtube client returned, so I formulated the request in the next cell to list the channel's first video of the year.

* Note the format of the datetime value is [International Standard ISO 8601](https://en.wikipedia.org/wiki/ISO_8601). You will certainly encounter other formats in your data science career, but I would be surprised if this is the last time you see this one.

start_date = "2023-01-02T00:00:00Z"
end_date = "2023-01-03T00:00:00Z"

request = youtube.search().list(
    part="snippet",
    channelId=CHANNEL_ID,
    publishedAfter=start_date,
    publishedBefore=end_date,
    type="video"
)
response = request.execute()

* Let's see what the response looks like.

response

* The response is a dictionary that includes lots of metadata about the video. 

* Now I'll figure out how to extract information that interests me. I'll stat with the title.

response['items'][0]['snippet']['title']

* And the description:

response['items'][0]['snippet']['description']

[Return to top](#Get-metadata-for-videos)

### Create a dataframe with channel metadata

I think I'm ready to dive in and get some data! First up is to identify the videos in January's "Effort" program. I happen to know there are twenty of them starting January 2 and ending January 28.

* Did someone say dataframe? My package of choice is Pandas. You might prefer something else. That's fine! Do what works for you.

import pandas as pd

* For this request, I specify `publishedBefore` and `publishedAfter` dates.

start_date = "2023-01-01T00:00:00Z"
end_date = "2023-01-30T00:00:00Z"

request = youtube.search().list(
    part="snippet",
    channelId=CHANNEL_ID,
    maxResults=31,
    publishedAfter=start_date,
    publishedBefore=end_date,
    type="video"
)
response = request.execute()

* We saw earlier that the response has lots of metadata! I invite you to explore the response beyond what I demonstrate here. In this case, I only care about the `items` key.

response.keys()

* I'll check that the response has the number of items I expect: twenty videos in the specified time frame.

len(response['items'])

* Great! Now I am ready to create the dataframe of metadata. I'll start with the video ids, titles, publication dates, and descriptions.

* You'll notice that I do a couple of checks after extracting the data: the length of the list and content of the first element in the list. 

vids = []
for item in response['items']:
    vid = item['id']['videoId']
    vids.append(vid)

len(vids)

vids[0]

titles = []
for item in response['items']:
    title = item['snippet']['title']
    titles.append(title)

len(titles)

titles[0]

times = []
for item in response['items']:
    time = item['snippet']['publishedAt']
    times.append(time)

len(times)

times[0]

* You may be thinking that this is getting a little repetitive and that I should write a function to get this done. If I were putting this into production somewhere, I would agree with you.

descs = []
for item in response['items']:
    desc = item['snippet']['description']
    descs.append(desc)

len(descs)

descs[0]

* We can now create the dataframe of metadata about Sydney's videos for the month of January.

data = pd.DataFrame({'vid':vids,
                     'title':titles,
                     'published':times,
                     'description':descs}
                   )

data.head()

* Looks good! I notice, however, that the `published` values are not in chronological order. This is probably going to bug me so I'll sort the dataframe by publication date.

* First convert to Panda's `datetime` format.

data['published'] = pd.to_datetime(data['published'])
data['published'].head()

* Now sort the dataframe by date.

data = data.sort_values('published')

data.head()

* That's better.

[Return to top](#Get-metadata-for-videos)

### Augment dataframe with video metadata

* To get more information about each video, we formulate a new request to access data about videos. 

def get_vid_stats(vid):
    request = youtube.videos().list(
        part="statistics",
        id=vid
    )
    response = request.execute()
    return response

* Let's see what kind of information we get about one of the videos in our dataframe.

test_vid = data.loc[0,'vid']
response = get_vid_stats(test_vid)
response

* I spy `statistics` in the `items` list. Now I'll figure out how to extract them.

* `response[items]` is a one-element list that is a dictionary. The key I am interested in is `statistics`.

statistics = response['items'][0]['statistics']
statistics

`statistics` itself is a dictionary with keys. Check them out:

keys = list(response['items'][0]['statistics'].keys())
keys

* All this information helps me formulate how I will extract statistical information from the response and add it to my data frame.

def extract_stats(row):
    '''
    Add video statistics to the row (a pd.Series)
    '''
    entry = row.copy()
    vid = entry['vid']
    response = get_vid_stats(vid)
    for key in keys:
        entry[key] = response['items'][0]['statistics'][key]
    return entry

* And test it out...

test_vid = data.loc[0]
extract_stats(test_vid)

* It works as expected, so I apply the function to all the rows to extract the statistics. Notice the `axis=1` parameter in the `apply` function because the function operates on a row. 

data = data.apply(extract_stats, axis=1)

* See what the dataframe looks like now...

data.head()

* At this point, my data science brain is popcorning with hypotheses about which day of the week has the most plays, if there are time series trends, how this January compares to last January...but I'm trying to stay on topic and will complete the task at hand: getting metadata!

* I'd like to figure out how long each work out is. Conveniently for my purposes, the answer is in the title of each video.

* I'll use a regular expression to extract the minutes. My favorite resource for testing out regex strings is [regex101.com](https://regex101.com). If you use this resource, don't put any sensitive information in the test string!!!

import re

* Test this out...I expect to get the number 30.

match = r'(\d+)'
matches = re.search(match, data.loc[0,'title'])
matches.groups()[0]

* Subtle but important to note ... this `'30'` is a string and not an integer. I'll make sure to convert it before storing in the dataframe.

def get_workout_time(title):
    '''
    The length of each workout in minutes is the first number in the title.
    Use a regular expression to get that value.
    '''
    match = r'(\d+)'
    matches = re.search(match, title)
    try:
        return int(matches.groups()[0])
    except:
        return -1

* Here's an example of applying a function to a column. I do not need to specificy `axis=0` because it is the default value. See the [documentation](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.apply.html) for a deeper explanation.

data['time'] = data['title'].apply(get_workout_time)

* Do a quick check:

data[['title','time']].head()

* Are you following along with a channel of your own interest? Look for a video's total time with a different request. Consult the YouTube Data API. You should find `duration` in the `part` called `contentDetails`. 

* YouTube has its own time format `PTXMYS` (where `X` and `Y` are numbers) so you'll want to formulate your own regular expression to extract the information you're looking for.

* The last bit of metadata that I want to extract is the date from the `published` column. 

* Conveniently, Pandas has this kind of conversion built in. If you can remember the syntax, more power to you. I have to google it every single time.

* Earlier, we converted the `published` column to Pandas `datetime` data type. To access different values of a `datetime`, use the `.dt` accessor. [See documentation for a deeper dive.](https://pandas.pydata.org/docs/reference/api/pandas.Series.dt.html?highlight=dt) 

* Now the last steps to polish this dataframe up: extract the publication date of each video, and then use these dates as the index. 

data['date'] = data['published'].dt.date

data = data.set_index('date')
data.head()

* Before I save this dataframe, I'm going to drop a couple of columns that don't look useful.

drop = ['published','favoriteCount']

data = data.drop(columns=drop)

data.head()

[Return to top](#Get-metadata-for-videos)

### Save the data

* Make sure to save the dataframe so you can use it in the next notebook, in which we will augment the dataframe with transcripts from each video and extract the motivational segment. 

data.to_csv('../data/2023-01-effort-metadata.csv')

[Return to top](#Get-metadata-for-videos)
