==> notebooks/02-transcripts.ipynb <==
# Video Transcripts

The main goal for this pet project is to capture Sydney's motivational talks. To do that, I need access to the transcripts for each video. This notebook captures that process.

Note: Section links work in JupyterLab but do not work when viewed in Github.

1. [Load](#Load-video-ids) video ids from the output of the [previous notebook](https://github.com/ltran17/motivational_messages/blob/main/notebooks/01-metadata.ipynb)
2. [Acquire](#Acquire-transcripts-for-each-video) transcripts for each video
3. [Extract](#Extract-the-motivational-talk-from-each-transcript) the motivational talk from each transcript
4. [Save](#Save-dataframe-to-a-file) dataframe to a file 

### Load video ids

import pandas as pd

I'm going to take a little side-trip here and talk about data hygiene.

* Raw data should never be modified. Results of data cleaning or other transformations should never overwrite the original file. Best practices for data pipelines sometimes refer to [Bronze, Silver, and Gold level data](https://www.linkedin.com/pulse/consider-gold-silver-bronze-your-data-just-olympics-ruaidhri-hallinan/). These are code-words for data in raw, interim, or processed stages.

* This is a small enough project that I'm not going to create separate folders for each stage of data. For any project larger than this, I would use the [Cookie Cutter Data Science file structure](https://drivendata.github.io/cookiecutter-data-science/#directory-structure).

* **GitHub is a code repository and not a data repository.** I do not push data to GitHub. Data versioning tools exist! I encourage you to look into those. 

All that is to say: you will not find the output of the previous notebook in this repo. You will need to run the code yourself and save the data where you want it to live.

* Here is where I stored the data:

datafile = '../data/2023-01-effort-metadata.csv'

* Since I know what my data looks like, I will take advantage of the parameters in the Pandas [`.read_csv` function](https://pandas.pydata.org/docs/reference/api/pandas.read_csv.html).

data = pd.read_csv(datafile, parse_dates=['date'], index_col='date')
data.head()

* And now I have the video ids.

vids = data['vid'].copy()
test_vid = vids[0]
test_vid

[Return to top](#Get-transcripts-for-each-video)

### Acquire transcripts for each video

The YouTube API does not make it easy to get the transcripts from videos! As I googled around to figure it out, I landed on [YouTube Transcript API](https://pypi.org/project/youtube-transcript-api/) (available through pypi.org). 

* This package makes the transcript acquisition task straightforward. I encourage you to familiarize yourself with the package through the link above or [fork it yourself from the GitHub page](https://github.com/jdepoix/youtube-transcript-api).

* **Caveat emptor**: The package author notes, "This code uses an undocumented part of the YouTube API, which is called by the YouTube web-client. So there is no guarantee that it won't stop working tomorrow, if they change how things work." Good to know.

from youtube_transcript_api import YouTubeTranscriptApi

* I'm going to test out YouTube Transcript API's `.get_transcript` function.

transcript = YouTubeTranscriptApi.get_transcript(test_vid)

type(transcript)

len(transcript)

* That's a long list! I'll inspect just the first few elements.

transcript[:3]

* `transcript` is a list of dictionaries with three keys each: the `text`, `start` and `duration` in seconds (to the hundredth place).

* I could write a function to extract the text values ... but the YouTubeTranscript API has already done this! How did I know it was there? I read [the documentation](https://github.com/jdepoix/youtube-transcript-api#using-formatters). This is a thoughtfully created package.

from youtube_transcript_api.formatters import TextFormatter

formatter = TextFormatter()
text = formatter.format_transcript(transcript)

type(text)

len(text)

text[:100]

* This formatter takes each of the `text` values from the `transcript` and appends them together with newline whitespace, `\n`, in between. I'll remove those.

text = text.replace('\n', ' ')

text[:100]

* Now that I know how to get the transcript for each video, I'll write a function.

* Because this is not part of the official YouTube API, I am going to be extra careful with these calls and not hammer the server with transcript requests. Maybe I'm being excessively cautious but I really don't want my YouTube access throttled.

import time

def get_transcript(vid, sleep=3):
    '''
    Returns formatted transcript for vid 
    Pause for sleep seconds before returning the transcript for vid
    '''
    transcript = YouTubeTranscriptApi.get_transcript(vid)
    text = formatter.format_transcript(transcript)
    text = text.replace('\n', ' ')
    time.sleep(3)
    return text

* Because of the pause in the function, this next cell takes a minute to run. If you are braver than I am, you can use a smaller value for the `sleep` parameter.

transcripts = vids.apply(get_transcript)

* Just how long are these transcripts? Too long to read through them all 😬

transcripts.str.len()

[Return to top](#Get-transcripts-for-each-video)

### Extract the motivational talk from each transcript

* Because I have done (literally) hundreds of Sydney's workouts, I know that the motivational talk almost always begins after the phrase, "You have made it to your cool down!"

* I explored the transcripts and discovered that sometimes this phrase is captured as "cool down" and sometimes as "cooldown." 

* There are also occasions where Sydney uses a different phrase, like "As the clock winds down to zero," or "Before you leave this workout."

phrases = ['to your cool down',
           'to your cooldown',
           'as the clock',
           'leave this workout']

def get_motivational_text(transcript):
    '''
    Return the motivational part of the transcript.
    '''
    text = transcript.lower()
    for phrase in phrases:
        loc = text.find(phrase)
        if loc > 0:
            motivation = text[loc+len(phrase):]
            return motivation
    return 'closing phrase not found'

Test out the function:

test_transcript = transcripts[0]
len(test_transcript)

motivation = get_motivational_text(test_transcript)
len(motivation)

Looks like it worked, but just to be sure...I will give it a read.

motivation

* That's still a lot of words! Many of them are boilerplate filler or directional -- "subscribe to the channel", "give this video a big thumbs up", "bend your knees slowly".

* This is good enough for me. I'm going to leave it to ChatGPT to filter out what I'm looking for.

* Time to get all the motivations.

motivations = transcripts.apply(get_motivational_text)

motivations

* Hmm...The motivation for January 17 is missing!

* Here's where you, data scientist, get to decide how to deal with missing data. How important is it to have this particular day's motivation? Do you want to impute the value, or can you drop the record?

* My choice is to drop the record. There's enough motivation to go around with the other 19 videos.

* Prepare the dataframe.

data = pd.DataFrame({'vid':vids,
              'motivation':motivations
             })
data

* I could hard code dropping the record for January 17 ... but future me will be really annoyed by this when I want motivations from, say, May 2024.

* My solution is to define a function to use as a filter.

def has_motivation(motivation):
    if motivation == 'closing phrase not found':
        return False
    return True

filt = data['motivation'].apply(has_motivation)
data = data[filt]
data

[Return to top](#Get-transcripts-for-each-video)

### Save dataframe to a file 

* This is ready to save! In the next notebook, I will ask ChatGPT to summarize these motivational talks.

data.to_csv('../data/2023-01-effort-motivations.csv')

[Return to top](#Get-transcripts-for-each-video)
