==> notebooks/03-motivation.ipynb <==
# Motivational Talks

I conceived of this project the same week as ChatGPT3 exploded into conversations at work, online, and at home. While my original plan had been to practice my NLP skills, I switched my goal to learn how to use this new tool. 

Note: Section links work in JupyterLab but do not work when this notebook is viewed in GitHub.

1. [Load](#Load-transcripts) transcripts from the output of the Notebook 02
2. [Interpret](#Interpret-response-from-ChatGPT3) response from ChatGPT3
3. [Play](#Play-with-the-prompt) with the prompt to get satisfactory response
4. [Summarize](#Summarize-motivational-talks) motivational talks for each video
5. [Save](#Save-the-motivational-messages) the motivational messages
6. [Consider](#Next-steps) next steps

### Load transcripts

I did the work of extracting the motivational part of each video's transcripts in the previous notebook. 

* I do not push data to code repositories, and encourage you to do the same. To keep everything organized on my machine, I have a `data` folder in my project and include this folder in the project's `.gitignore` file.

filename = '../data/2023-01-effort-motivations.csv'

import pandas as pd

data = pd.read_csv(filename, parse_dates=['date'])
data.head()

[Return to top](#Extract-motivational-messages)

### Interpret response from ChatGPT3

* ChatGPT3 has multiple language models available for use. It looks like [Davinci](https://platform.openai.com/docs/models/davinci) will be the right one for this task of interpreting and summarizing.

* Different models will give different results. The creators of an endless, AI-generated, Seinfeld parody [learned this the hard way](https://wapo.st/3I7BSHW) when they switched from Davinci to Curie.

* For this project, I am not worried about generating hateful or harming content. I am taking note of [OpenAI's moderation tools](https://platform.openai.com/docs/guides/moderation/overview) for future work.

* The next three cells are set up.

with open('../config/chat_gpt3_api') as file:
    CHAT_KEY = file.readline()

* Even more important than not pushing data to a code repository is to **never push a key to a code repository**! I have stored the keys in a folder called `config` and added this folder to the project's `.gitignore` file. There are other ways to accomplish this, choose your preferred method.

import openai
openai.api_key = CHAT_KEY
models = openai.Model.list()

test_text = data.loc[0,'motivation']
test_text[:100]

* The template for the function in the next cell comes from OpenAI. If you used the playground when you were setting up API access, you already encountered the `temperature` and `max_tokens` parameters.

* I interpret `temperature` as a creativity setting. Set to 0, multiple calls with the same text will return similar (if not identical) responses. Set to 1, the model gets very creative! 

* You can find no better explanation of OpenAI's tokens than the [documentation](https://help.openai.com/en/articles/4936856-what-are-tokens-and-how-to-count-them) from the source. If you want a more general discussion of tokens and tokenization, send "nlp tokens" through your favorite search engine to learn more.

def get_response(text):
    prompt = 'extract the motivational message from the following text and summarize in three sentences: '
    response = openai.Completion.create(
        model='text-davinci-003',
        prompt=prompt+text,
        temperature=0.7,
        max_tokens=25
    )
    return response

response = get_response(test_text)
response

* For a first pass, this is a pretty good summary! Let me see how to extract it from the response.

response['choices'][0]['text']

[Return to top](#Extract-motivational-messages)

### Play with the prompt

* I notice that it cut off the summary -- no doubt because I limited the `max_tokens` to 25.

* I also notice that it really sounds like a summary! I would rather it sound like a motivational message. I need to adjust the prompt.

def get_response(text):
    prompt = 'extract the motivational message from the following text in three sentences: '
    response = openai.Completion.create(
        model='text-davinci-003',
        prompt=prompt+text,
        temperature=0.7,
        max_tokens=100
    )
    return response

response = get_response(test_text)
response['choices'][0]['text']

* This sounds better! However, the model is interpreting some of the boilerplate text as motivational. Adjust the prompt further.

def get_response(text):
    directions = 'extract the motivational message from the following text in three sentences: '
    conditions = 'Ignore sentences that direct viewers to subscribe, like, buy, or comment.'
    response = openai.Completion.create(
        model='text-davinci-003',
        prompt=directions+text+conditions,
        temperature=0.7,
        max_tokens=100
    )
    return response

response = get_response(test_text)
response['choices'][0]['text']

* Not bad! This sounds almost like Sydney herself. I am going to move on with the `get_response` function as defined above.

* Before I get the motivations from all of the videos, I'll write a couple of functions streamline this process.

def extract_text(response):
    '''
    Return the text from the response object
    '''
    text = response['choices'][0]['text']
    return text

* Also, I will be polite by pausing between requests. 

import time

def get_motivation(text, sleep_time=3):
    '''
    Return the extracted motivation from the text, pausing for sleep_time seconds after request.
    '''
    response = get_response(text)
    motivation = extract_text(response)
    time.sleep(sleep_time)
    return motivation

* Test this out...

get_motivation(test_text)

[Return to top](#Extract-motivational-messages)

### Summarize motivational talks

* Alright! I am ready to get the motivations for the month! 

* This next call will take a minute because of (1) the sleep parameter and (2) however busy the ChatGPT3 servers are. In fact...the first time I ran the next cell, it returned a rate limit error. That's what happens when the continental US is awake and pinging the servers at 11 am Eastern. I had more success at 4 pm.

motivations = data['motivation'].apply(get_motivation)
motivations[0]

* Motivation verging on platitudes. It's ok! 

motivations

* I see that my prompt to create a motivational message resulted in multiple texts with the word "motivation." In future iterations, I might edit the prompt.

* I also notice inconsistent newlines -- some text was returned with this whitespace, and some was not. I'm going to clean that up.

motivations = motivations.str.strip()

for motivation in motivations.head():
    print(motivation+'\n\n')

* That is ... a lot. What happens if I whittle these down to one sentence each?

* I'll add a parameter for the number of sentences. (If I ever clean up this notebook for production code, there will be some serious refactoring.)

def get_response(text, sentences):
    '''
    Get response from ChatGPT with length that is the number of sentences
    '''
    directions = f'extract the motivational message from the following text in {sentences} sentences: '
    conditions = 'Ignore sentences that direct viewers to subscribe, like, buy, or comment.'
    response = openai.Completion.create(
        model='text-davinci-003',
        prompt=directions+text+conditions,
        temperature=0.7,
        max_tokens=100
    )
    return response

def get_motivation(text, sentences=3, sleep_time=3):
    '''
    Return the extracted motivation from the text in number of sentences, 
    pausing for sleep_time seconds after request.
    '''
    response = get_response(text, sentences)
    motivation = extract_text(response)
    time.sleep(sleep_time)
    return motivation

* Because I added a parameter, I have to use a lambda function to apply `get_motivation` to the series.

shorter_motivations = motivations.apply(lambda m: get_motivation(m, 1))
shorter_motivations[0]

* I see the whitespace issue is showing up again. Will remove...

shorter_motivations = shorter_motivations.str.strip()
for s in shorter_motivations.head():
    print(s+'\n')

* This is better! I am reasonably satisfied with these summaries.

* One last thing I want to do is summarize the entire month of motivations into one, just to see what happens. Start by combining all of the responses into one, long text.

all_motivations = ' '.join(shorter_motivations)

all_motivations

* Now use that as the text to prompt ChatGPT3.

month_motivation = get_motivation(all_motivations, sentences=3)
month_motivation

[Return to top](#Extract-motivational-messages)

### Save the motivational messages

* Ok, there we have it. Even though ChatGPT did not *quite* follow directions with this four-sentence summary, its responses reflect the month's theme of Effort.

* Even though I don't know quite what to do with these summaries, I'm going to save them for future use.

summaries = data[['date','vid']]
summaries['summary'] = shorter_motivations
summaries.head()

summaries.to_csv('../data/2023-01-effort-summaries.csv')

with open('../data/2023-01-effort-summary.txt', 'w') as file:
    file.write(month_motivation.strip())

[Return to top](#Extract-motivational-messages)

### Next steps

* I might transcribe the monthly summary into my workout journal.

* I might clean up the code to turn these notebooks into a script that I can run at the end of each month.

* I could go back to exploring statistics from Notebook 1 -- gather more data from the channel and create some visualizations.

* The YouTube channel metadata included urls for images. I could automate creation of a workout journal that includes pictures, workout descriptions, and motivational talks.

* It's also ok if there are no next steps! **I met my goal of getting my head around these new-to-me APIs and testing out the ChatGPT3 technology.**

[Return to top](#Extract-motivational-messages)
